fix(datainsightsconfig): Add aria label to group form controls
When grouped form controls have an accessible name, it helps a screen
reader user to make sense of a form. The accessible name is read out to
the user and describes the purpose of related controls such as a set of
options, credit card details, or an address.

If the form is long or complicated, it also helps to distinguish one
group of form controls from another.